Remove old winternitz files, edit winternitz_hash.rs accrodingly, bug fix, refactor and some changes for the ease of usage
