+++ src/signatures/winternitz_hash.rs
@@ -1,11 +1,12 @@
-use crate::treepp::*;
-use crate::signatures::winternitz::{sign, checksig_verify, generate_public_key, PublicKey
-};
+ use crate::treepp::*;
+use super::winternitz::*;
 use crate::hash::blake3::blake3_160_var_length;
 use blake3::hash;
 
-const MESSAGE_HASH_LEN: u8 = 20;
-
+const MESSAGE_HASH_LEN: u32 = 20;
+//These two are just added for compatibility, they can be changed and might not be the best choice of algorithms or parameters for your usage
+static O: Winternitz::<ListpickVerifier, StraightforwardConverter> = Winternitz::new();
+static PS: Parameters = Parameters::new(MESSAGE_HASH_LEN * 2, 4);
 
 /// Verify a Winternitz signature for the hash of the top `input_len` many bytes on the stack
 /// The hash function is blake3 with a 20-byte digest size
@@ -13,7 +14,7 @@ const MESSAGE_HASH_LEN: u8 = 20;
 pub fn check_hash_sig(public_key: &PublicKey, input_len: usize) -> Script {
     script! {
         // 1. Verify the signature and compute the signed message
-        { checksig_verify(&public_key) }
+        { O.checksig_verify(&PS, &public_key) }
         for _ in 0..MESSAGE_HASH_LEN {
             OP_TOALTSTACK
         }
@@ -34,32 +35,35 @@ pub fn check_hash_sig(public_key: &PublicKey, input_len: usize) -> Script {
 }
 
 /// Create a Winternitz signature for the blake3 hash of a given message
-pub fn sign_hash(sec_key: &str, message: &[u8]) -> Script {
+pub fn sign_hash(sec_key: &Vec<u8>, message: &[u8]) -> Script {
     let message_hash = hash(message);
     let message_hash_bytes = &message_hash.as_bytes()[0..20];
     script! {
-        { sign(sec_key, message_hash_bytes) }
+        { O.sign(&PS, sec_key, &message_hash_bytes.to_vec()) }
     }
 }
 
 #[cfg(test)]
 mod test {
     use super::*;
+    const MY_SEC_KEY: &str = "b138982ce17ac813d505b5b40b665d404e9528e7";
 
     #[test]
     fn test_check_hash_sig() {
-
         // My secret key 
-        let my_sec_key = "b138982ce17ac813d505b5b40b665d404e9528e7";
+        let secret_key = match hex::decode(MY_SEC_KEY) {
+            Ok(bytes) => bytes,
+            Err(_) => panic!("Invalid hex string"),
+        };
         
         // My public key
-        let public_key = generate_public_key(my_sec_key);
+        let public_key = generate_public_key(&PS, &secret_key);
 
         // The message to sign
         let message = *b"This is an arbitrary length input intended for testing purposes....";
 
 
-        run(script! {
+        assert!(execute_script(script! {
             //
             // Unlocking Script
             //
@@ -69,7 +73,7 @@ mod test {
                 { *byte }
             }
             // 2. Push the signature
-            { sign_hash(my_sec_key, &message) }
+            { sign_hash(&secret_key, &message) }
             
             
             //
@@ -77,7 +81,7 @@ mod test {
             //
             { check_hash_sig(&public_key, message.len()) }
             OP_TRUE
-        });   
+        }).success == true);   
     }
 
 }
